Extract basic consent manager example into its own file
